realy respect -D in all cases
git-svn-id: http://svn.berlios.de/svnroot/repos/sipsak/trunk@433 75b5f7c7-cfd4-0310-b54c-e118b2c5249a
(cherry picked from commit f5681475477f964ed4a87e42b5a08ab686cb07d0)
(cherry picked from commit 66fafcfdb84fd39fc475c8bff9117e2ad789dba0)

+++ transport.c
@@ -805,7 +805,7 @@ int check_for_message(char *recv, int size, struct sipsak_con_data *cd,
 			}
 		}
 		senddiff = deltaT(&(srt->starttime), &(srt->recvtime));
-		if (senddiff > inv_final) {
+		if (senddiff > (float)inv_final) {
 			if (timing == 0) {
 				if (verbose>0)
 					printf("*** giving up, no final response after %.3f ms\n", senddiff);
